+++ packages/test-environment/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ricsam/isolate-test-environment
 
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @ricsam/isolate-core@0.1.17
+
 ## 0.1.16
 
 ### Patch Changes